+++ tests/ProductSearchTest.php
@@ -10,39 +10,34 @@ class ProductSearchTest extends \PHPUnit\Framework\TestCase
 
     public function testSetup()
     {
-        $testIp = '45.248.78.5';
+        $testIp = '173.92.179.208';
         $testAgent = 'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_13_5) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/67.0.3396.99 Safari/537.36';
 
-        $baseify = (new Baseify('28a0fd38a28698b4788fcd882544bd16'))->productSearch([
-            'domain' => 'discount-savings.com'
+        $baseify = (new Baseify('ba7c6b686c25e6ac8bddefc948d40d4e'))->productSearch([
+            'domain' => 'discount-savings.com',
+            'widget' => 'test'
         ]);
 
         $baseify->user()->set('ip',$testIp);
         $baseify->user()->set('ua',$testAgent);
-        $baseify->filter()->set('limit',999);
+        $baseify->filter()->set('limit',99);
 
         $results = $baseify->query('Matte Lipcolor');
 
         // print_r($results->getEndpoint());
-
         // print_r($results->getRaw());
 
-        $status = $results->getStatus();
-        $products = $results->getProducts();
-
-        $stores = $results->getStores();
-        // print_r($stores);
-
-        $brands = $results->getBrands();
-        // print_r($brands);
-
-        $categories = $results->getCategories();
-        // print_r($categories);
+        $status      = $results->getStatus();
+        $products    = $results->getProducts();
+        $stores      = $results->getStores();
+        $brands      = $results->getBrands();
+        $categories  = $results->getCategories();
+        $categoryIds = $results->getCategoryWithIds();
+        // print_r($categoryIds);
 
         print_r($products->all());
 
         $this->assertEquals(true, $status);
-        // $this->assertEquals(9, $products->count());
         $this->assertInternalType('object', $products);
         $this->assertInternalType('array', $products->all());
         $this->assertInstanceOf(Collection::class, $products);
@@ -55,7 +50,7 @@ class ProductSearchTest extends \PHPUnit\Framework\TestCase
         $testIp = '127.0.0.1';
         $testAgent = 'Test';
 
-        $baseify = (new Baseify('28a0fd38a28698b4788fcd882544bd16'))->productSearch();
+        $baseify = (new Baseify('ba7c6b686c25e6ac8bddefc948d40d4e'))->productSearch();
 
         $baseify->user()->set('ip',$testIp);
         $baseify->user()->set('ua',$testAgent);